+++ src/src/App.vue
@@ -2,59 +2,77 @@
   <v-app>
     <v-app-bar
       app
-      color="primary"
-      dark
     >
-      <div class="d-flex align-center">
-        <v-img
-          alt="Vuetify Logo"
-          class="shrink mr-2"
-          contain
-          src="https://cdn.vuetifyjs.com/images/logos/vuetify-logo-dark.png"
-          transition="scale-transition"
-          width="40"
-        />
-
-        <v-img
-          alt="Vuetify Name"
-          class="shrink mt-1 hidden-sm-and-down"
-          contain
-          min-width="100"
-          src="https://cdn.vuetifyjs.com/images/logos/vuetify-name-dark.png"
-          width="100"
-        />
-      </div>
-
-      <v-spacer></v-spacer>
-
-      <v-btn
-        href="https://github.com/vuetifyjs/vuetify/releases/latest"
-        target="_blank"
-        text
+      <v-app-bar-nav-icon
+        class="d-md-none"
+        @click="drawer = true"
+      />
+      <v-toolbar-title class="my-title">
+        yaso28 - Web系エンジニア
+      </v-toolbar-title>
+      <v-tabs
+        right
+        class="d-none d-md-block"
       >
-        <span class="mr-2">Latest Release</span>
-        <v-icon>mdi-open-in-new</v-icon>
-      </v-btn>
+        <v-tab
+          v-for="menuItem in menuItemList"
+          :key="menuItem"
+          :to="{name: menuItem}"
+          exact
+        >
+          {{ menuItem }}
+        </v-tab>
+      </v-tabs>
     </v-app-bar>
 
+    <v-navigation-drawer
+      v-model="drawer"
+      fixed
+      temporary
+    >
+      <v-list
+        nav
+        dense
+      >
+        <v-list-item-group>
+          <v-list-item
+            v-for="menuItem in menuItemList"
+            :key="menuItem"
+            :to="{name: menuItem}"
+            exact
+          >
+            <v-list-item-title>{{ menuItem }}</v-list-item-title>
+          </v-list-item>
+        </v-list-item-group>
+      </v-list>
+    </v-navigation-drawer>
+
     <v-main>
-      <HelloWorld/>
+      <v-container>
+        <router-view />
+      </v-container>
     </v-main>
   </v-app>
 </template>
 
 <script>
-import HelloWorld from './components/HelloWorld'
-
 export default {
   name: 'App',
-
-  components: {
-    HelloWorld
+  data () {
+    return {
+      drawer: false
+    }
   },
-
-  data: () => ({
-    //
-  })
+  computed: {
+    menuItemList () {
+      return ['HOME', 'ABOUT', 'SKILLS', 'WORKS']
+    }
+  }
 }
 </script>
+
+<style lang="scss">
+.v-toolbar__title.my-title {
+  overflow: visible;
+}
+</style>