+++ src/src/consts/index.js
@@ -1,0 +1,3 @@
+export default Object.freeze({
+  site_name: 'yaso28 - Web系エンジニア'
+})