+++ src/src/main.js
@@ -2,6 +2,9 @@ import Vue from 'vue'
 import App from './App.vue'
 import router from './router'
 import vuetify from './plugins/vuetify'
+import VueMeta from 'vue-meta'
+
+Vue.use(VueMeta)
 
 Vue.config.productionTip = false