+++ src/src/router/index.js
@@ -1,22 +1,28 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
-import Home from '../views/Home.vue'
 
 Vue.use(VueRouter)
 
 const routes = [
   {
     path: '/',
-    name: 'Home',
-    component: Home
+    name: 'HOME',
+    component: () => import('@/views/Home')
   },
   {
     path: '/about',
-    name: 'About',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
+    name: 'ABOUT',
+    component: () => import('@/views/About')
+  },
+  {
+    path: '/skills',
+    name: 'SKILLS',
+    component: () => import('@/views/Skills')
+  },
+  {
+    path: '/works',
+    name: 'WORKS',
+    component: () => import('@/views/Works')
   }
 ]