+++ src/src/views/Home.vue
@@ -1,18 +1,25 @@
 <template>
-  <div class="home">
-    <img alt="Vue logo" src="../assets/logo.png">
-    <HelloWorld msg="Welcome to Your Vue.js App"/>
+  <div>
+    <h1>HOME</h1>
   </div>
 </template>
 
 <script>
-// @ is an alias to /src
-import HelloWorld from '@/components/HelloWorld.vue'
+import consts from '@/consts'
 
 export default {
   name: 'Home',
-  components: {
-    HelloWorld
+  metaInfo () {
+    return {
+      title: consts.site_name,
+      meta: [
+        {
+          vmid: 'description',
+          name: 'description',
+          content: 'Web系エンジニアyaso28のポートフォリオサイトです。'
+        }
+      ]
+    }
   }
 }
 </script>